[FIX] fix the front code

+++ console/views/app_config/app_label.py
@@ -115,7 +115,9 @@ class AppLabelView(AppBaseView):
             label_id = request.data.get("label_id", None)
             if not label_id:
                 return Response(general_message(400, "param error", "标签ID未指定"), status=400)
-
+            service_label = service_label_repo.get_service_label(label_id)
+            if not service_label:
+                return Response(general_message(400, "tag does not exist", "标签不存在"), status=400)
             code, msg, event = label_service.delete_service_label(self.tenant, self.service, label_id)
             if code != 200:
                 return Response(general_message(code, "add labels error", msg), status=code)
